seperate calulator.py and autofocus on calculator
